Correzioni per la compatibilità con windows

+++ matlab/lib/Image.cpp
@@ -156,11 +156,11 @@ void Image::integralImage(Image *origin, Image *destination) {
 void Image::scaleImage(Image *origin, Image *destination, unsigned int scale) {
     double value;
     try {
-        for (unsigned int j = scale-1, y = 0;
-             j < origin->height;
+        for (unsigned int j = scale - 1, y = 0;
+             j < (unsigned int) origin->height;
              j += scale, y++) {
-            for (unsigned int i = scale-1, x = 0;
-                 i < origin->width;
+            for (unsigned int i = scale - 1, x = 0;
+                 i < (unsigned int) origin->width;
                  i += scale, x++) {
                 // Scrittura dei pixel nella nuova immagine riscalata
                 value = origin->read(i, j);
